+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - parsing <think> tags as reasoning part of the response
 - fixed unnecessary reload of component during onboarding process
 - implemented better support for themes + new themes added
+- implemented support for fonts + new fonts added
 
 ## [0.26.0]